move set.go to internal and address comments
