+++ internal/container/set/set_test.go
@@ -1,4 +1,4 @@
-package main
+package set
 
 import "testing"
